protorh/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
docutils==0.20.1
fastapi==0.104.1
h11==0.14.0
idna==3.4
lockfile==0.12.2
pid==3.0.4
Pillow==10.1.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-daemon==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
service==0.6.0
services==0.1.1
setproctitle==1.3.3
sniffio==1.3.0
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
